// pc/src/services/apis/models/KvPairVoString.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 企业全生命周期
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface KvPairVoString
 */
export interface KvPairVoString {
    /**
     * 
     * @type {string}
     * @memberof KvPairVoString
     */
    value: string;
    /**
     * 
     * @type {string}
     * @memberof KvPairVoString
     */
    label: string;
}


/**
 * Check if a given object implements the KvPairVoString interface.
 */
export function instanceOfKvPairVoString(value: object): value is KvPairVoString {
    if (!('value' in value) || value['value'] === undefined) return false;
    if (!('label' in value) || value['label'] === undefined) return false;
    return true;
}

export function KvPairVoStringFromJSON(json: any): KvPairVoString {
    return KvPairVoStringFromJSONTyped(json, false);
}

export function KvPairVoStringFromJSONTyped(json: any, ignoreDiscriminator: boolean): KvPairVoString {
    if (json == null) {
        return json;
    }
    return {
        
        'value': json['value'],
        'label': json['label'],
    };
}

export function KvPairVoStringToJSON(json: any): KvPairVoString {
    return KvPairVoStringToJSONTyped(json, false);
}

export function KvPairVoStringToJSONTyped(value?: KvPairVoString | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'value': value['value'],
        'label': value['label'],
    };
}