// pc/src/services/apis/models/MenuVo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 企业全生命周期
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MenuVo
 */
export interface MenuVo {
    /**
     * 
     * @type {string}
     * @memberof MenuVo
     */
    id?: string;
    /**
     * 页面路径
     * @type {string}
     * @memberof MenuVo
     */
    path?: string;
    /**
     * 目录名称
     * @type {string}
     * @memberof MenuVo
     */
    name: string;
    /**
     * 目录图标
     * @type {string}
     * @memberof MenuVo
     */
    icon: string;
    /**
     * 排序
     * @type {number}
     * @memberof MenuVo
     */
    sort: number;
    /**
     * 父级菜单
     * @type {string}
     * @memberof MenuVo
     */
    parentId: string;
    /**
     * 客户端
     * @type {string}
     * @memberof MenuVo
     */
    endpoint: MenuVoEndpointEnum;
    /**
     * 子菜单
     * @type {Array<MenuVo>}
     * @memberof MenuVo
     */
    children: Array<MenuVo>;
    /**
     * Alias For 'id'
     * @type {string}
     * @memberof MenuVo
     */
    value?: string;
    /**
     * Alias For 'name'
     * @type {string}
     * @memberof MenuVo
     */
    title: string;
}


/**
 * @export
 */
export const MenuVoEndpointEnum = {
    Pc: 'PC',
    H5: 'H5'
} as const;
export type MenuVoEndpointEnum = typeof MenuVoEndpointEnum[keyof typeof MenuVoEndpointEnum];



/**
 * Check if a given object implements the MenuVo interface.
 */
export function instanceOfMenuVo(value: object): value is MenuVo {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('icon' in value) || value['icon'] === undefined) return false;
    if (!('sort' in value) || value['sort'] === undefined) return false;
    if (!('parentId' in value) || value['parentId'] === undefined) return false;
    if (!('endpoint' in value) || value['endpoint'] === undefined) return false;
    if (!('children' in value) || value['children'] === undefined) return false;
    if (!('title' in value) || value['title'] === undefined) return false;
    return true;
}

export function MenuVoFromJSON(json: any): MenuVo {
    return MenuVoFromJSONTyped(json, false);
}

export function MenuVoFromJSONTyped(json: any, ignoreDiscriminator: boolean): MenuVo {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'path': json['path'] == null ? undefined : json['path'],
        'name': json['name'],
        'icon': json['icon'],
        'sort': json['sort'],
        'parentId': json['parentId'],
        'endpoint': json['endpoint'],
        'children': ((json['children'] as Array<any>).map(MenuVoFromJSON)),
        'value': json['value'] == null ? undefined : json['value'],
        'title': json['title'],
    };
}

export function MenuVoToJSON(json: any): MenuVo {
    return MenuVoToJSONTyped(json, false);
}

export function MenuVoToJSONTyped(value?: MenuVo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'path': value['path'],
        'name': value['name'],
        'icon': value['icon'],
        'sort': value['sort'],
        'parentId': value['parentId'],
        'endpoint': value['endpoint'],
        'children': ((value['children'] as Array<any>).map(MenuVoToJSON)),
        'value': value['value'],
        'title': value['title'],
    };
}